# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * LivingSyria API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorEnvelope {
  error: string;
}

export interface AuthUser {
  id: string;
  /** @nullable */
  email: string | null;
  /** @nullable */
  firstName: string | null;
  /** @nullable */
  lastName: string | null;
  /** @nullable */
  profileImageUrl: string | null;
}

export interface AuthUserEnvelope {
  user: AuthUser | null;
}

export interface MobileTokenExchangeRequest {
  /** @minLength 1 */
  code: string;
  /** @minLength 1 */
  code_verifier: string;
  /** @minLength 1 */
  redirect_uri: string;
  /** @minLength 1 */
  state: string;
  /** @minLength 1 */
  nonce?: string;
}

export interface MobileTokenExchangeSuccess {
  token: string;
}

export const LogoutSuccessValue = {
  success: true,
} as const;
export type LogoutSuccess = typeof LogoutSuccessValue;

export interface UploadUrlRequest {
  /** @minLength 1 */
  name: string;
  /** @minimum 1 */
  size: number;
  /** @minLength 1 */
  contentType: string;
}

export interface UploadUrlResponse {
  uploadURL: string;
  objectPath: string;
  metadata?: UploadUrlRequest;
}

export interface Category {
  id: string;
  slug: string;
  nameAr: string;
  /** @nullable */
  nameEn?: string | null;
  /** @nullable */
  icon?: string | null;
  /** @nullable */
  parentId?: string | null;
  sortOrder: number;
}

export interface Listing {
  id: string;
  slug: string;
  userId: string;
  categoryId: string;
  titleAr: string;
  /** @nullable */
  titleEn?: string | null;
  descriptionAr: string;
  /** @nullable */
  descriptionEn?: string | null;
  /** @nullable */
  priceCents?: number | null;
  currency: string;
  isFree: boolean;
  isNegotiable: boolean;
  country: string;
  city: string;
  /** @nullable */
  district?: string | null;
  status: string;
  listingType: string;
  viewCount: number;
  /** @nullable */
  primaryImageUrl?: string | null;
  /** @nullable */
  publishedAt?: string | null;
  createdAt: string;
}

export interface ListingImage {
  id: string;
  listingId: string;
  objectPath: string;
  sortOrder: number;
  isPrimary: boolean;
}

export type ListingDetail = Listing & {
  images: ListingImage[];
  category: Category;
};

export interface PageMeta {
  page: number;
  limit: number;
  total: number;
  pages: number;
}

export interface ListingsPage {
  data: Listing[];
  meta: PageMeta;
}

export type CreateListingBodyCurrency =
  (typeof CreateListingBodyCurrency)[keyof typeof CreateListingBodyCurrency];

export const CreateListingBodyCurrency = {
  USD: "USD",
  SYP: "SYP",
  EUR: "EUR",
} as const;

export interface CreateListingBody {
  categoryId: string;
  /** @minLength 3 */
  titleAr: string;
  titleEn?: string;
  /** @minLength 10 */
  descriptionAr: string;
  descriptionEn?: string;
  /** @minimum 0 */
  priceCents?: number;
  currency?: CreateListingBodyCurrency;
  isFree?: boolean;
  isNegotiable?: boolean;
  country?: string;
  city: string;
  district?: string;
  imageObjectPaths?: string[];
}

export interface NewsArticle {
  id: string;
  slug: string;
  sourceName: string;
  sourceUrl: string;
  titleAr: string;
  /** @nullable */
  titleEn?: string | null;
  /** @nullable */
  summaryAr?: string | null;
  /** @nullable */
  summaryEn?: string | null;
  /** @nullable */
  bodyAr?: string | null;
  /** @nullable */
  bodyEn?: string | null;
  /** @nullable */
  aiSummaryAr?: string | null;
  /** @nullable */
  aiSummaryEn?: string | null;
  /** @nullable */
  coverImageUrl?: string | null;
  tags: string[];
  status: string;
  viewCount: number;
  publishedAt: string;
  createdAt: string;
}

export interface NewsPage {
  data: NewsArticle[];
  meta: PageMeta;
}

/**
 * Opaque session token — `Bearer <sid>`.
 */
export type AuthorizationSessionHeaderParameter = string;

export type BeginBrowserLoginParams = {
  returnTo?: string;
};

export type HandleBrowserLoginCallbackParams = {
  code?: string;
  state?: string;
  iss?: string;
};

export type ListCategories200 = {
  data: Category[];
};

export type ListListingsParams = {
  category?: string;
  city?: string;
  q?: string;
  /**
   * @minimum 1
   */
  page?: number;
  /**
   * @minimum 1
   * @maximum 50
   */
  limit?: number;
};

export type CreateListing201 = {
  data: Listing;
};

export type ListMyListings200 = {
  data: Listing[];
};

export type GetListing200 = {
  data: ListingDetail;
};

export type ListNewsParams = {
  tag?: string;
  /**
   * @minimum 1
   */
  page?: number;
  /**
   * @minimum 1
   * @maximum 50
   */
  limit?: number;
};

export type GetNewsArticle200 = {
  data: NewsArticle;
};

export type TriggerNewsIngest200Data = {
  started: boolean;
  count: number;
};

export type TriggerNewsIngest200 = {
  data: TriggerNewsIngest200Data;
};