# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: LivingSyria API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: Auth
    description: Browser and mobile authentication endpoints.
  - name: Storage
    description: Object storage upload and serving endpoints.
  - name: Categories
    description: Listing category taxonomy.
  - name: Listings
    description: Classifieds listings.
  - name: News
    description: AI-summarized news articles.
  - name: Admin
    description: Admin-only operations.
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/user:
    get:
      tags: [Auth]
      operationId: getCurrentAuthUser
      summary: Get the currently authenticated user
      parameters:
        - $ref: "#/components/parameters/AuthorizationSessionHeader"
        - $ref: "#/components/parameters/SessionCookie"
      responses:
        "200":
          description: Auth status resolved successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUserEnvelope"
  /login:
    get:
      tags: [Auth]
      operationId: beginBrowserLogin
      summary: Start the browser OIDC login flow
      parameters:
        - name: returnTo
          in: query
          required: false
          schema:
            type: string
      responses:
        "302":
          description: Redirect to the OIDC authorization endpoint.
  /callback:
    get:
      tags: [Auth]
      operationId: handleBrowserLoginCallback
      summary: Complete the browser OIDC login flow
      parameters:
        - name: code
          in: query
          required: false
          schema:
            type: string
        - name: state
          in: query
          required: false
          schema:
            type: string
        - name: iss
          in: query
          required: false
          schema:
            type: string
            format: uri
      responses:
        "302":
          description: Redirect after callback.
  /logout:
    get:
      tags: [Auth]
      operationId: logoutBrowserSession
      summary: Clear the session and begin OIDC logout
      parameters:
        - $ref: "#/components/parameters/AuthorizationSessionHeader"
        - $ref: "#/components/parameters/SessionCookie"
      responses:
        "302":
          description: Redirect to OIDC logout.
  /mobile-auth/token-exchange:
    post:
      tags: [Auth]
      operationId: exchangeMobileAuthorizationCode
      summary: Exchange a mobile OIDC code for a session token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MobileTokenExchangeRequest"
      responses:
        "200":
          description: Session created successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MobileTokenExchangeSuccess"
        "400":
          description: Bad request.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "401":
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "500":
          description: Server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /mobile-auth/logout:
    post:
      tags: [Auth]
      operationId: logoutMobileSession
      summary: Delete a mobile session token
      parameters:
        - $ref: "#/components/parameters/AuthorizationSessionHeader"
      responses:
        "200":
          description: Logout processed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LogoutSuccess"

  /storage/uploads/request-url:
    post:
      tags: [Storage]
      operationId: requestUploadUrl
      summary: Request a presigned URL for file upload
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UploadUrlRequest"
      responses:
        "200":
          description: Presigned upload URL generated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadUrlResponse"
        "400":
          description: Bad request.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "500":
          description: Server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /storage/public-objects/{filePath}:
    get:
      tags: [Storage]
      operationId: getPublicObject
      summary: Serve a public asset
      parameters:
        - name: filePath
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Object content.
          content:
            "*/*":
              schema:
                type: string
                format: binary
        "404":
          description: Not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "500":
          description: Server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /storage/objects/{objectPath}:
    get:
      tags: [Storage]
      operationId: getStorageObject
      summary: Serve an object entity
      parameters:
        - name: objectPath
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Object content.
          content:
            "*/*":
              schema:
                type: string
                format: binary
        "404":
          description: Not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "500":
          description: Server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"

  /categories:
    get:
      tags: [Categories]
      operationId: listCategories
      summary: List all categories
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Category"

  /listings:
    get:
      tags: [Listings]
      operationId: listListings
      summary: List listings
      parameters:
        - name: category
          in: query
          required: false
          schema:
            type: string
        - name: city
          in: query
          required: false
          schema:
            type: string
        - name: q
          in: query
          required: false
          schema:
            type: string
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 20
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListingsPage"
    post:
      tags: [Listings]
      operationId: createListing
      summary: Create a new listing (auth required)
      parameters:
        - $ref: "#/components/parameters/AuthorizationSessionHeader"
        - $ref: "#/components/parameters/SessionCookie"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateListingBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    $ref: "#/components/schemas/Listing"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"

  /listings/me:
    get:
      tags: [Listings]
      operationId: listMyListings
      summary: List the authenticated user's listings
      parameters:
        - $ref: "#/components/parameters/AuthorizationSessionHeader"
        - $ref: "#/components/parameters/SessionCookie"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Listing"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"

  /listings/{slug}:
    get:
      tags: [Listings]
      operationId: getListing
      summary: Get a listing by slug
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    $ref: "#/components/schemas/ListingDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"

  /news:
    get:
      tags: [News]
      operationId: listNews
      summary: List published news articles
      parameters:
        - name: tag
          in: query
          required: false
          schema:
            type: string
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 12
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NewsPage"

  /news/{slug}:
    get:
      tags: [News]
      operationId: getNewsArticle
      summary: Get a news article by slug
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    $ref: "#/components/schemas/NewsArticle"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"

  /admin/news/ingest:
    post:
      tags: [Admin]
      operationId: triggerNewsIngest
      summary: Trigger an RSS ingestion + AI summarization run
      parameters:
        - $ref: "#/components/parameters/AuthorizationSessionHeader"
        - $ref: "#/components/parameters/SessionCookie"
      responses:
        "200":
          description: Started
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: object
                    required: [started, count]
                    properties:
                      started:
                        type: boolean
                      count:
                        type: integer
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"

components:
  parameters:
    AuthorizationSessionHeader:
      name: Authorization
      in: header
      required: false
      description: Opaque session token — `Bearer <sid>`.
      schema:
        type: string
    SessionCookie:
      name: sid
      in: cookie
      required: false
      description: Browser session cookie.
      schema:
        type: string
  schemas:
    HealthStatus:
      type: object
      required: [status]
      properties:
        status:
          type: string
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
    AuthUser:
      type: object
      required: [id, email, firstName, lastName, profileImageUrl]
      properties:
        id:
          type: string
        email:
          type: [string, "null"]
          format: email
        firstName:
          type: [string, "null"]
        lastName:
          type: [string, "null"]
        profileImageUrl:
          type: [string, "null"]
    AuthUserEnvelope:
      type: object
      required: [user]
      properties:
        user:
          oneOf:
            - $ref: "#/components/schemas/AuthUser"
            - type: "null"
    MobileTokenExchangeRequest:
      type: object
      required: [code, code_verifier, redirect_uri, state]
      properties:
        code:
          type: string
          minLength: 1
        code_verifier:
          type: string
          minLength: 1
        redirect_uri:
          type: string
          minLength: 1
          format: uri
        state:
          type: string
          minLength: 1
        nonce:
          type: string
          minLength: 1
    MobileTokenExchangeSuccess:
      type: object
      required: [token]
      properties:
        token:
          type: string
    LogoutSuccess:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
          const: true
    UploadUrlRequest:
      type: object
      required: [name, size, contentType]
      properties:
        name:
          type: string
          minLength: 1
        size:
          type: integer
          minimum: 1
        contentType:
          type: string
          minLength: 1
    UploadUrlResponse:
      type: object
      required: [uploadURL, objectPath]
      properties:
        uploadURL:
          type: string
          format: uri
        objectPath:
          type: string
        metadata:
          $ref: "#/components/schemas/UploadUrlRequest"
    Category:
      type: object
      required: [id, slug, nameAr, sortOrder]
      properties:
        id:
          type: string
        slug:
          type: string
        nameAr:
          type: string
        nameEn:
          type: [string, "null"]
        icon:
          type: [string, "null"]
        parentId:
          type: [string, "null"]
        sortOrder:
          type: integer
    Listing:
      type: object
      required:
        - id
        - slug
        - userId
        - categoryId
        - titleAr
        - descriptionAr
        - currency
        - isFree
        - isNegotiable
        - country
        - city
        - status
        - listingType
        - viewCount
        - createdAt
      properties:
        id:
          type: string
        slug:
          type: string
        userId:
          type: string
        categoryId:
          type: string
        titleAr:
          type: string
        titleEn:
          type: [string, "null"]
        descriptionAr:
          type: string
        descriptionEn:
          type: [string, "null"]
        priceCents:
          type: [integer, "null"]
        currency:
          type: string
        isFree:
          type: boolean
        isNegotiable:
          type: boolean
        country:
          type: string
        city:
          type: string
        district:
          type: [string, "null"]
        status:
          type: string
        listingType:
          type: string
        viewCount:
          type: integer
        primaryImageUrl:
          type: [string, "null"]
        publishedAt:
          type: [string, "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
    ListingImage:
      type: object
      required: [id, listingId, objectPath, sortOrder, isPrimary]
      properties:
        id:
          type: string
        listingId:
          type: string
        objectPath:
          type: string
        sortOrder:
          type: integer
        isPrimary:
          type: boolean
    ListingDetail:
      allOf:
        - $ref: "#/components/schemas/Listing"
        - type: object
          required: [images, category]
          properties:
            images:
              type: array
              items:
                $ref: "#/components/schemas/ListingImage"
            category:
              $ref: "#/components/schemas/Category"
    ListingsPage:
      type: object
      required: [data, meta]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Listing"
        meta:
          $ref: "#/components/schemas/PageMeta"
    PageMeta:
      type: object
      required: [page, limit, total, pages]
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        pages:
          type: integer
    CreateListingBody:
      type: object
      required: [categoryId, titleAr, descriptionAr, city]
      properties:
        categoryId:
          type: string
        titleAr:
          type: string
          minLength: 3
        titleEn:
          type: string
        descriptionAr:
          type: string
          minLength: 10
        descriptionEn:
          type: string
        priceCents:
          type: integer
          minimum: 0
        currency:
          type: string
          enum: [USD, SYP, EUR]
        isFree:
          type: boolean
        isNegotiable:
          type: boolean
        country:
          type: string
        city:
          type: string
        district:
          type: string
        imageObjectPaths:
          type: array
          items:
            type: string
    NewsArticle:
      type: object
      required:
        - id
        - slug
        - sourceName
        - sourceUrl
        - titleAr
        - tags
        - status
        - viewCount
        - publishedAt
        - createdAt
      properties:
        id:
          type: string
        slug:
          type: string
        sourceName:
          type: string
        sourceUrl:
          type: string
        titleAr:
          type: string
        titleEn:
          type: [string, "null"]
        summaryAr:
          type: [string, "null"]
        summaryEn:
          type: [string, "null"]
        bodyAr:
          type: [string, "null"]
        bodyEn:
          type: [string, "null"]
        aiSummaryAr:
          type: [string, "null"]
        aiSummaryEn:
          type: [string, "null"]
        coverImageUrl:
          type: [string, "null"]
        tags:
          type: array
          items:
            type: string
        status:
          type: string
        viewCount:
          type: integer
        publishedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
    NewsPage:
      type: object
      required: [data, meta]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/NewsArticle"
        meta:
          $ref: "#/components/schemas/PageMeta"