# lib/api-zod/src/generated/types/listingsPage.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * LivingSyria API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Listing } from "./listing";
import type { PageMeta } from "./pageMeta";

export interface ListingsPage {
  data: Listing[];
  meta: PageMeta;
}