# lib/api-zod/src/generated/types/newsArticle.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * LivingSyria API specification
 * OpenAPI spec version: 0.1.0
 */

export interface NewsArticle {
  id: string;
  slug: string;
  sourceName: string;
  sourceUrl: string;
  titleAr: string;
  /** @nullable */
  titleEn?: string | null;
  /** @nullable */
  summaryAr?: string | null;
  /** @nullable */
  summaryEn?: string | null;
  /** @nullable */
  bodyAr?: string | null;
  /** @nullable */
  bodyEn?: string | null;
  /** @nullable */
  aiSummaryAr?: string | null;
  /** @nullable */
  aiSummaryEn?: string | null;
  /** @nullable */
  coverImageUrl?: string | null;
  tags: string[];
  status: string;
  viewCount: number;
  publishedAt: Date;
  createdAt: Date;
}