# lib/api-zod/src/generated/types/pageMeta.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * LivingSyria API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PageMeta {
  page: number;
  limit: number;
  total: number;
  pages: number;
}